[libmfx_core_vaapi] Use derived image directly to improve performance
This is just a work around to trigger the performance improvement.

Currently, in libmfx core library, sw copy is used to copy derived data
to system memory. This can be redundant under some situations since the
derived data could be used directly (like sample decode).

If the frame size is large, the performance can be greatly influenced.

Signed-off-by: Linjie Fu <linjie.fu@intel.com>